run: use IP instead of INET module for supporting IPv6
Some servers have the ipaddr ban list for certain IPv4 addresses, which,
unfortunately, sometimes (when the user uses dynamic addr) are blocked but the
user has nothing to do with that and, therefore, the bot is also blocked.

This patch adds IO::Socket::IP module for handling both IPv4 and v6 upon user
configuration, in case he wants to force it.

Signed-off-by: Bruno Meneguele <bmeneg@redhat.com>