We still need to evaluate the 001, always!
In the event that the server sends 001 as its first message, if
the {user} setup hasn't happened, that happens, and the later elsif
that picks up the 001 never happens--so, your bot won't join channels
listed in connect( join=> [] ), or execute commands in run(). This
remedies that, by allowing analysis of the line to happen immediately
after setting up the user, and not just tossing the line and going
to the next.