Eval 001 regardless of return location
In the event the server sends 001 as its first message, if the user
setup hasn't happened, that happens, and the later elsif that picks up
the 001 never happens, which results in the bot not joining channels
listed in connect( join=> [] ) and not executing commands in run().

This patch fixes this by allowing analysis of the line to happen
immediately after setting up the user.

(Thank you to GeekRuthie.)